requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
PyJWT
passlib
pydantic
pydantic_settings
psycopg2-binary